requirements.txt
python-dotenv
chainlit
google-generativeai
